# ENH: speed up einsum with optimize using batched matmul