This PR enables dispatching *all* pairwise calls in `einsum` when `optimize=` is turned on to batched matrix multiply (or non-batched `matmul` and `multiply` as special cases). For a short explanation of the implementation see here - https://github.com/jcmgray/einsum_bmm.

Currently falling back to `c_einsum` for these, when there exist batch indices, induces potentially many orders of magnitude slow downs (#22604, e.g. 7000x), particularly in the case of large or high dimensional tensors.

After fairly extensive benchmarking (below), I *believe* that this should be a uniform improvement across essentially all the cases when one would use `optimize=True`. This is partly enabled by the modern performance of matmul which in the past was not always faster than `einsum`. However, I know `einsum` has an enormous number of uses and I'd gladly accept more suggestions for benchmarks, and would understand people's hesitation! 

(@dgasmith, @seberg)

## Other notes:
- This replaces all the `can_blas` logic and `tensordot` calls, and simplifies everything into the `bmm_einsum`, which encapsulates how `tensordot` is implemented
- When there are no contracted indices, `numpy.multiply` is used for slightly better performance
- The current implementation can simply pass `out` and other kwargs on to `mutiply` and `matmul`, unlike tensordot, meaning `einsum('ij,jk->ik', x, y, optimize=True, out=z)` should write directly to `z`
- The fusing allows `einsum(eq, ..., optimize=True)` to be used when there are more than 32 indices involved (#5744).
- The low overhead of the actuall `bmm_einsum` implementation comes from caching what operations to use based on `(eq, x.shape, y.shape)`, using `@functools.lru_cache(2**12)`.
- the performance broadly speaking is brought in line with `torch.einsum`

## TODO: 
- [ ] update docs?
- [ ] decide cache size?
- [x] should I lint the existing `einsumfunc.py` code?

## Benchmarks

In the following the kernels are the following:

- `no optimize`: the base `np.einsum(eq, x, y)` i.e. `c_einsum`
- `optimize + dot`: the current `np.einsum(eq, x, y, optimize=True)` which calls `tensordot` where possible, and also induces some overhead from the potential path optimization (even though there are only two terms in the following, the equation must be checked and parsed etc.)
- `optimize + bmm`: the proposed `np.einsum(eq, x, y, optimize=True)` which calls batched matmul, and still induces some overhead from the potential path optimization
- `no optimize + bmm`: this calls the pairwise bmm einsum directly, and is provided simply to show what overhead comes from the bmm impl, and what from the potential path optimization 

The main two to compare are `optimize + dot` to `optimize + bmm`. Overall I'd summarize the results as the following:

1. when a contraction is memory bound, all methods are pretty much equivalent
2. when a contraction involves no batch dimensions, performance of `dot` is retained
3. when a contraction involves batch dimensions, performance can be orders of magnitude faster

All dimensions are size `n` unless otherwise noted.

<img width="1328" height="650" alt="vec_inner" src="https://github.com/user-attachments/assets/2b90de50-0cfc-403c-944c-627710332941" />

<img width="1328" height="650" alt="vec_outer" src="https://github.com/user-attachments/assets/bc955303-7475-4a96-8b39-030046ad8133" />

<img width="1328" height="650" alt="square_matmul" src="https://github.com/user-attachments/assets/a81aa244-fab4-4e6d-ab61-3623a4e649a0" />

<img width="1328" height="650" alt="square_matvec" src="https://github.com/user-attachments/assets/531d1481-53f7-4431-aaf5-98cd6782be52" />

<img width="1328" height="650" alt="square_vecmat" src="https://github.com/user-attachments/assets/53644f4e-bfd9-4ea4-9b8b-75d9edc36ede" />

For `batch_matmul_small` the sizes are `(n, 2, 2), (n, 2, 2)`.

<img width="1328" height="650" alt="batch_matmul_small" src="https://github.com/user-attachments/assets/15d43a70-a3b9-4ca9-8d4c-c048a9f44dfd" />

For `batch_matmul_equal` the sizes are `(n, n, n), (n, n, n)`.

<img width="1328" height="650" alt="batch_matmul_equal" src="https://github.com/user-attachments/assets/7ebb5070-2681-4c51-8200-32a96c23f53d" />

For `batch_matmul_large` the sizes are `(10, n, n), (10, n, n)`.

<img width="1328" height="650" alt="batch_matmul_large" src="https://github.com/user-attachments/assets/8d5b8cb4-0594-4f0c-af73-e1c7ee6c62a6" />

<img width="1328" height="650" alt="hadamard" src="https://github.com/user-attachments/assets/6fbf5cca-6dff-490c-881e-92eefbf2ddc7" />

<img width="1328" height="650" alt="hadamard_unalinged" src="https://github.com/user-attachments/assets/e477d602-c0f4-43f6-b57f-bb4f3d4c0c1f" />

<img width="1328" height="650" alt="CCSDT_1" src="https://github.com/user-attachments/assets/b6bf94df-228e-4585-bd3f-c5c7b7412123" />

<img width="1328" height="650" alt="interleaved_dot" src="https://github.com/user-attachments/assets/a0632f67-9f5c-4ceb-9ec7-d014d800e479" />

<img width="1328" height="650" alt="interleaved_batched_dot" src="https://github.com/user-attachments/assets/f7d54568-4ac9-42ee-81f8-06778c06f074" />


In `random_extreme` the shapes are taken as the following:
```python
[
    (n, 4, 3, 4, 3, 4, 2, 4, 2, n, 2, n, 3, 2, 4),
    (2, 4, n, n, 4, 4, 4, 3, 4, 4, 4, 3, 4)
]
```

<img width="1406" height="648" alt="random_extreme" src="https://github.com/user-attachments/assets/e2197ea1-bf9e-44c6-a992-cfbb8c110fa5" />


In `many_small_dims_dot` the shapes are `([2] * n, [2] * n)` and half in an interleaved pattern are contracted.

<img width="1328" height="648" alt="many_small_dims_dot" src="https://github.com/user-attachments/assets/37b01048-9c4e-4dba-9f00-cf071901969a" />

In `many_small_dims_batched_dot` the shapes are `([2] * n, [2] * n)` and a third in an interleaved pattern are contracted and a third in an interleaved pattern are batch indices.

<img width="1328" height="648" alt="many_small_dims_batched_dot" src="https://github.com/user-attachments/assets/32eba347-37d4-49b9-bf0c-67aaabfea0b0" />