@jcmgray 2023-05-17
Hi @seberg, ah yes I just checked and currently that matmul behaviour would result in a performance regression for the specific case that arrays reach the matmul call (after preparatory reshape/transpose etc) with whichever strides trigger the non-blas, no-copy call (#23588).

```python
import numpy as np
d = 1000
x = np.random.randn(d, d)
y = np.random.randn(d, d) + 1j * np.random.randn(d, d)

%%timeit
z = np.einsum('ij,jk->ik', x, y.real, optimize=True)
# -->
# 11.5 ms ± 473 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  (current dot backend)
# -->
# 897 ms ± 23.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  (new matmul backend)
```

If the solution in #23752 is to do the copy when needed then that is very much aligned with the logic of `optimize=True`, i.e. trade memory and copying for being able to BLAS-it. Else one could check in the einsum_bmm call but preferably it would be deferred to `matmul` I suppose.