@dgasmith 2023-05-30
> TBH, I had hoped @dgasmith can give a recommendation. The timings do look like this should work pretty generically, although its easy to miss cases, I guess.

It is hard for me to say, we benchmarked fairly throughly with `optimize=True`; however, the deployed result had a large number of edge cases that we missed. That being said, localizing the code to `optimize=True` is a good move to let everyone switch on/off performance.