@jcmgray 2023-07-26
Rather than wait on progress in https://github.com/numpy/numpy/issues/23588 would it be reasonable to simply require contiguity before the matmul in the `bmm_einsum` function here? 

I.e. do we expect there to be cases where the arrays are not contiguous, and we expect it to be faster to leave them so?

My guess is that this is only the case for certain small contractions that are not the focus of when people use `optimize=True`. 

Of course having some rough heuristic in `matmul` itself would be cleaner.