@jcmgray 2025-09-17
As hoped, with https://github.com/numpy/numpy/issues/23588 the following no longer has any slowdown (and any logic about inputs contiguity is deferred to `matmul`):
```python
import numpy as np
d = 1000
x = np.random.randn(d, d)
y = np.random.randn(d, d) + 1j * np.random.randn(d, d)

%%timeit
z = np.einsum('ij,jk->ik', x, y.real, optimize=True)
```
I have also updated all the benchmarks above - the results are pretty much the same. From my perspective this is ready, but still happy to add more test contractions.