@jcmgray 2025-11-05
Great, I'll add a algorithmic note. 

---

I thought I'd also give some extra analysis of the speedups that might be useful for future changes or discussion. The question is basically whether the (rare) cases where the bmm approach is slower that `c_einsum` (other than very small sizes) can be systematically identified. One thing you can look at is arithmetic intensity, here defined by `intensity = prod(dimensions_of_all_involved_indices) / (size_a + size_b + size_output)`. 

- high intensity = matmul like
- low intensity = outer / inner / hadamard product like

If you have a batched matmul with dimensions `(B i j) (B j k) -> (B i k)` then intensity is maximized when `i == j == k`.

If you compare that with speedup with a 'soft cutoff' of 1ms `speedup = (time_einsum + 0.001) / (time_bmm + 0.001)` (since we don't care too much about very small contractions) and bin the same set of contractions above you get a plot like this:

![Untitled](https://github.com/user-attachments/assets/8fa86bac-e3da-4173-b152-259ecda4f464)

In other words:
- intensity is a pretty big predictor for speedup
- all the cases which perform worse have low intensity

It might be tempting to special case low intensity, however, if you plot the mean speedup:

![Untitled](https://github.com/user-attachments/assets/572a9ea6-0813-4747-80ff-b48d0e07ed2d)

Then it is always (just) above 1, so the weight of cases where *low intensity is still sped up by the bmm approach* just about wins out (and includes important cases like `eq = "Bij,Bjk->Bik"` with i, j and k equal but small).

So for the moment, I don't think are super easy ways to customize the dispatch logic. One would probably also have to think about secondary factors like the cost of transposing and reshaping.