@mattip 2025-11-12
Something changed in the latest iteration, the `bench_linalg.Einsum.time_einsum_noncon_contig_contig(<class 'numpy.float64'>)` benchmark is now slower than `main`. Is this expected?
```
| Change   | Before [3abf5795] <einsum_bmm~2>   | After [d2c1ccfc] <einsum_bmm>   |   Ratio | Benchmark (Parameter)                                                             |
|----------|------------------------------------|---------------------------------|---------|-----------------------------------------------------------------------------------|
| +        | 16.9±0.2μs                         | 17.8±0.2μs                      |    1.06 | bench_linalg.Einsum.time_einsum_noncon_contig_contig(<class 'numpy.float64'>)     |
| -        | 11.0±0.2ms                         | 10.4±0.2ms                      |    0.95 | bench_linalg.MatmulStrided.time_matmul('matmul_m400_p400_n100_bs01')              |
| -        | 3.12±0.04μs                        | 2.94±0.06μs                     |    0.94 | bench_linalg.LinalgSmallArrays.time_det_small_array                               |
| -        | 28.7±0.1μs                         | 26.5±0.3μs                      |    0.92 | bench_linalg.Einsum.time_einsum_multiply(<class 'numpy.float32'>)                 |
| -        | 25.5±0.2μs                         | 23.4±0.1μs                      |    0.92 | bench_linalg.Einsum.time_einsum_sum_mul2(<class 'numpy.float32'>)                 |
| -        | 28.8±0.07μs                        | 26.2±0.1μs                      |    0.91 | bench_linalg.Einsum.time_einsum_noncon_multiply(<class 'numpy.float32'>)          |
| -        | 33.5±0.2μs                         | 30.4±0.5μs                      |    0.91 | bench_linalg.Einsum.time_einsum_noncon_multiply(<class 'numpy.float64'>)          |
| -        | 25.3±0.3μs                         | 23.0±0.6μs                      |    0.91 | bench_linalg.Einsum.time_einsum_sum_mul(<class 'numpy.float32'>)                  |
| -        | 33.6±0.2μs                         | 30.2±0.8μs                      |    0.9  | bench_linalg.Einsum.time_einsum_multiply(<class 'numpy.float64'>)                 |
| -        | 66.9±0.2μs                         | 59.5±1μs                        |    0.89 | bench_linalg.Einsum.time_einsum_contig_outstride0(<class 'numpy.float64'>)        |
| -        | 7.25±0.5ms                         | 6.49±0.05ms                     |    0.89 | bench_linalg.Linalg.time_pinv('complex128')                                       |
| -        | 1.51±0.05ms                        | 1.33±0.01ms                     |    0.88 | bench_linalg.Einsum.time_einsum_noncon_outer(<class 'numpy.float32'>)             |
| -        | 17.7±0.4μs                         | 15.0±0.1μs                      |    0.85 | bench_linalg.Einsum.time_einsum_noncon_mul(<class 'numpy.float32'>)               |
| -        | 16.5±0.03μs                        | 14.0±0.1μs                      |    0.85 | bench_linalg.Einsum.time_einsum_noncon_mul(<class 'numpy.float64'>)               |
| -        | 40.9±0.1μs                         | 34.2±0.5μs                      |    0.84 | bench_linalg.Einsum.time_einsum_contig_outstride0(<class 'numpy.float32'>)        |
| -        | 28.6±0.1μs                         | 23.8±0.7μs                      |    0.83 | bench_linalg.Einsum.time_einsum_noncon_sum_mul(<class 'numpy.float32'>)           |
| -        | 29.1±0.05μs                        | 24.1±0.2μs                      |    0.83 | bench_linalg.Einsum.time_einsum_noncon_sum_mul2(<class 'numpy.float32'>)          |
| -        | 56.2±0.2μs                         | 43.5±0.7μs                      |    0.77 | bench_linalg.Einsum.time_einsum_contig_contig(<class 'numpy.float32'>)            |
| -        | 264±10μs                           | 173±8μs                         |    0.65 | bench_linalg.Einsum.time_einsum_mul(<class 'numpy.float32'>)                      |
| -        | 8.11±0.3ms                         | 5.10±0.2ms                      |    0.63 | bench_linalg.Einsum.time_einsum_noncon_outer(<class 'numpy.float64'>)             |
| -        | 97.7±1μs                           | 58.6±1μs                        |    0.6  | bench_linalg.Einsum.time_einsum_contig_contig(<class 'numpy.float64'>)            |
| -        | 12.2±0.1μs                         | 7.34±0.2μs                      |    0.6  | bench_linalg.Einsum.time_einsum_noncon_contig_outstride0(<class 'numpy.float32'>) |
| -        | 12.2±0.1μs                         | 7.13±0.2μs                      |    0.58 | bench_linalg.Einsum.time_einsum_noncon_contig_outstride0(<class 'numpy.float64'>) |
| -        | 10.6±0.3ms                         | 5.95±0.05ms                     |    0.56 | bench_linalg.Einsum.time_einsum_outer(<class 'numpy.float32'>)                    |
| -        | 22.5±0.4ms                         | 12.7±0.09ms                     |    0.56 | bench_linalg.Einsum.time_einsum_outer(<class 'numpy.float64'>)                    |
| -        | 184±20μs                           | 102±5μs                         |    0.55 | bench_linalg.Einsum.time_einsum_mul(<class 'numpy.float64'>)                      |

```